refactor: split the cipher module and move them to the separate folder
